(PC-21736)[API] refactor: Move requests helpers to zendesk_sell backend
