+++ api/.env.testing
@@ -85,6 +85,6 @@ SMS_NOTIFICATION_BACKEND=pcapi.notifications.sms.backends.sendinblue.ToDevSendin
 WEBAPP_V2_URL=https://app.testing.passculture.team
 WEBAPP_V2_REDIRECT_URL=https://redirect.testing.passculture.team
 ZENDESK_SELL_API_URL=https://api.getbase.com
-ZENDESK_SELL_BACKEND=pcapi.core.external.backends.testing.TestingBackend
+ZENDESK_SELL_BACKEND=pcapi.core.external.backends.logger.LoggerBackend
 BIG_QUERY_NOTIFICATIONS_TABLE_BASENAME="passculture-data-ehp.backend_dev"
 SESSION_COOKIE_SAMESITE="None"
